fix: 픽스존 댓글 삭제 휴지통 이미지로 변경
